Always extract a sync token from sync API response
When retrieving a page from the sync API the response will contain
a 'nextPageUrl' if there are multiple pages of results or a 'nextSyncUrl'
if the current page is the last in the set of results.

ResourceBuilder.build was only handling cases where the result was the
last page in a set of results (it only returned a SyncPage when it saw a
'nextSyncUrl' in the response JSON). This meant that for a result set of
multiple pages ResourceBuilder.build incorrectly returned an Array
object instead of a SyncPage.

The SyncPage code was also only able to extract a sync token from a
response that was the last page in a set of results (i.e., it contained
a 'nextSyncUrl').

This commit adds the property SyncPage.next_page_url and modifies
SyncPage._get_next_token to extract a sync token from either
the 'nextPageUrl' or 'nextSyncUrl' (whichever is present).
ResourceBuilder now checks for both 'nextPageUrl' and 'nextSyncUrl' and
returns a SyncPage.